refactor(export): move rewrite_refs next to the other line surgery
`plan.rs` crossed the 400-line house rule at 406 when the previous commit added
`next_offset`, and the module-size ratchet caught it in CI. Splitting rather
than allowlisting, per AGENTS.md.

`rewrite_refs` moves to `line_edit.rs`, which already holds every other
operation on the `#N` references of a single STEP line (`arg_refs`,
`classify_refs`, `decide_line`); its module doc widens from the container drop
to line surgery generally. Its test moves with it. Pure relocation — no
behaviour change, and the merge of duplex.ifc + AC20-FZK-Haus.ifc still drops
the same 17 containers with 0 dangling references.

plan.rs 406 -> 364, line_edit.rs 217 -> 263; every merged/ module is now back
under the limit with no new allowlist row.

Ref #3643.
