+++ rust/export/src/merged/plan_tests.rs
@@ -30,20 +30,6 @@ fn resolve_included_pulls_forward_closure() {
 }
 
 #[test]
-fn rewrite_refs_offsets_and_redirects() {
-    let mut remap = HashMap::new();
-    remap.insert(2u32, 100u32);
-    let out = rewrite_refs(
-        b"#3=IFCRELAGGREGATES('r #2',$,$,$,#1,(#2));",
-        10,
-        &|n| remap.get(&n).copied(),
-    );
-    // #1 offset by 10 → #11; #2 redirected → #100; '#2' in the string untouched.
-    assert!(out.contains("#11,(#100))"));
-    assert!(out.contains("'r #2'"));
-}
-
-#[test]
 fn redundant_rel_aggregates_dropped_only_when_fully_shared() {
     let idx = ModelIndex::build(TWO_STOREYS.as_bytes());
     let mut skip = HashSet::new();
